backend/requirements.txt
Flask
flask-cors
supabase
tensorflow==2.12
opencv-python
numpy
Pillow
PyJWT
